pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-dance>=7.1.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "flask-login>=0.6.3",
    "oauthlib>=3.2.2",
    "flask-babel>=4.0.0",
    "sqlalchemy>=2.0.41",
    "werkzeug>=3.1.3",
    "pillow>=11.2.1",
    "googlemaps>=4.10.0",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
